+++ codex-rs/core/src/agent/control_tests.rs
@@ -7,11 +7,12 @@ use crate::config::AgentRoleConfig;
 use crate::config::Config;
 use crate::config::ConfigBuilder;
 use crate::config_loader::LoaderOverrides;
-use crate::contextual_user_message::SUBAGENT_NOTIFICATION_OPEN_TAG;
 use crate::features::Feature;
+use crate::model_visible_context::SUBAGENT_NOTIFICATION_OPEN_TAG;
 use assert_matches::assert_matches;
 use codex_protocol::config_types::ModeKind;
 use codex_protocol::models::ContentItem;
+use codex_protocol::models::MessageRole;
 use codex_protocol::models::ResponseItem;
 use codex_protocol::protocol::ErrorEvent;
 use codex_protocol::protocol::EventMsg;
@@ -97,7 +98,7 @@ fn has_subagent_notification(history_items: &[ResponseItem]) -> bool {
         let ResponseItem::Message { role, content, .. } = item else {
             return false;
         };
-        if role != "user" {
+        if role != "developer" {
             return false;
         }
         content.iter().any(|content_item| match content_item {
@@ -381,7 +382,7 @@ async fn spawn_agent_can_fork_parent_thread_history() {
     let harness = AgentControlHarness::new().await;
     let (parent_thread_id, parent_thread) = harness.start_thread().await;
     parent_thread
-        .inject_user_message_without_turn("parent seed context".to_string())
+        .inject_message_without_turn(MessageRole::User, "parent seed context".to_string())
         .await;
     let turn_context = parent_thread.codex.session.new_default_turn().await;
     let parent_spawn_call_id = "spawn-call-history".to_string();
