+++ codex-rs/core/src/plugins/render.rs
@@ -88,12 +88,5 @@ pub(crate) fn render_explicit_plugin_instructions(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn render_plugin_instructions_returns_none_for_empty_plugins() {
-        assert_eq!(render_plugin_instructions(&[]), None);
-    }
-}
+#[path = "render_tests.rs"]
+mod tests;
