+++ codex-rs/core/src/plugins/render_tests.rs
@@ -2,8 +2,8 @@ use super::*;
 use pretty_assertions::assert_eq;
 
 #[test]
-fn render_plugins_section_returns_none_for_empty_plugins() {
-    assert_eq!(render_plugins_section(&[]), None);
+fn render_plugin_instructions_returns_none_for_empty_plugins() {
+    assert_eq!(render_plugin_instructions(&[]), None);
 }
 
 #[test]
