+++ codex-rs/core/src/project_doc_tests.rs
@@ -83,7 +83,7 @@ fn load_test_skills(config: &Config) -> crate::skills::SkillLoadOutcome {
 async fn no_doc_file_returns_none() {
     let tmp = tempfile::tempdir().expect("tempdir");
 
-    let res = get_user_instructions(&make_config(&tmp, 4096, None).await, None, None).await;
+    let res = get_user_instructions(&make_config(&tmp, 4096, None).await, None).await;
     assert!(
         res.is_none(),
         "Expected None when AGENTS.md is absent and no system instructions provided"
@@ -97,7 +97,7 @@ async fn doc_smaller_than_limit_is_returned() {
     let tmp = tempfile::tempdir().expect("tempdir");
     fs::write(tmp.path().join("AGENTS.md"), "hello world").unwrap();
 
-    let res = get_user_instructions(&make_config(&tmp, 4096, None).await, None, None)
+    let res = get_user_instructions(&make_config(&tmp, 4096, None).await, None)
         .await
         .expect("doc expected");
 
@@ -116,7 +116,7 @@ async fn doc_larger_than_limit_is_truncated() {
     let huge = "A".repeat(LIMIT * 2); // 2 KiB
     fs::write(tmp.path().join("AGENTS.md"), &huge).unwrap();
 
-    let res = get_user_instructions(&make_config(&tmp, LIMIT, None).await, None, None)
+    let res = get_user_instructions(&make_config(&tmp, LIMIT, None).await, None)
         .await
         .expect("doc expected");
 
@@ -148,7 +148,7 @@ async fn finds_doc_in_repo_root() {
     let mut cfg = make_config(&repo, 4096, None).await;
     cfg.cwd = nested;
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("doc expected");
     assert_eq!(res, "root level doc");
@@ -160,7 +160,7 @@ async fn zero_byte_limit_disables_docs() {
     let tmp = tempfile::tempdir().expect("tempdir");
     fs::write(tmp.path().join("AGENTS.md"), "something").unwrap();
 
-    let res = get_user_instructions(&make_config(&tmp, 0, None).await, None, None).await;
+    let res = get_user_instructions(&make_config(&tmp, 0, None).await, None).await;
     assert!(
         res.is_none(),
         "With limit 0 the function should return None"
@@ -175,7 +175,7 @@ async fn js_repl_instructions_are_appended_when_enabled() {
         .enable(Feature::JsRepl)
         .expect("test config should allow js_repl");
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("js_repl instructions expected");
     let expected = "## JavaScript REPL (Node)\n- Use `js_repl` for Node-backed JavaScript with top-level await in a persistent kernel.\n- `js_repl` is a freeform/custom tool. Direct `js_repl` calls must send raw JavaScript tool input (optionally with first-line `// codex-js-repl: timeout_ms=15000`). Do not wrap code in JSON (for example `{\"code\":\"...\"}`), quotes, or markdown code fences.\n- Helpers: `codex.cwd`, `codex.homeDir`, `codex.tmpDir`, `codex.tool(name, args?)`, and `codex.emitImage(imageLike)`.\n- `codex.tool` executes a normal tool call and resolves to the raw tool output object. Use it for shell and non-shell tools alike. Nested tool outputs stay inside JavaScript unless you emit them explicitly.\n- `codex.emitImage(...)` adds one image to the outer `js_repl` function output each time you call it, so you can call it multiple times to emit multiple images. It accepts a data URL, a single `input_image` item, an object like `{ bytes, mimeType }`, or a raw tool response object with exactly one image and no text. It rejects mixed text-and-image content.\n- `codex.tool(...)` and `codex.emitImage(...)` keep stable helper identities across cells. Saved references and persisted objects can reuse them in later cells, but async callbacks that fire after a cell finishes still fail because no exec is active.\n- Request full-resolution image processing with `detail: \"original\"` only when the `view_image` tool schema includes a `detail` argument. The same availability applies to `codex.emitImage(...)`: if `view_image.detail` is present, you may also pass `detail: \"original\"` there. Use this when high-fidelity image perception or precise localization is needed, especially for CUA agents.\n- Example of sharing an in-memory Playwright screenshot: `await codex.emitImage({ bytes: await page.screenshot({ type: \"jpeg\", quality: 85 }), mimeType: \"image/jpeg\", detail: \"original\" })`.\n- Example of sharing a local image tool result: `await codex.emitImage(codex.tool(\"view_image\", { path: \"/absolute/path\", detail: \"original\" }))`.\n- When encoding an image to send with `codex.emitImage(...)` or `view_image`, prefer JPEG at about 85 quality when lossy compression is acceptable; use PNG when transparency or lossless detail matters. Smaller uploads are faster and less likely to hit size limits.\n- Top-level bindings persist across cells. If a cell throws, prior bindings remain available and bindings that finished initializing before the throw often remain usable in later cells. For code you plan to reuse across cells, prefer declaring or assigning it in direct top-level statements before operations that might throw. If you hit `SyntaxError: Identifier 'x' has already been declared`, first reuse the existing binding, reassign a previously declared `let`, or pick a new descriptive name. Use `{ ... }` only for a short temporary block when you specifically need local scratch names; do not wrap an entire cell in block scope if you want those names reusable later. Reset the kernel with `js_repl_reset` only when you need a clean state.\n- Top-level static import declarations (for example `import x from \"./file.js\"`) are currently unsupported in `js_repl`; use dynamic imports with `await import(\"pkg\")`, `await import(\"./file.js\")`, or `await import(\"/abs/path/file.mjs\")` instead. Imported local files must be ESM `.js`/`.mjs` files and run in the same REPL VM context. Bare package imports always resolve from REPL-global search roots (`CODEX_JS_REPL_NODE_MODULE_DIRS`, then cwd), not relative to the imported file location. Local files may statically import only other local relative/absolute/`file://` `.js`/`.mjs` files; package and builtin imports from local files must stay dynamic. `import.meta.resolve()` returns importable strings such as `file://...`, bare package names, and `node:...` specifiers. Local file modules reload between execs, while top-level bindings persist until `js_repl_reset`.\n- Avoid direct access to `process.stdout` / `process.stderr` / `process.stdin`; it can corrupt the JSON line protocol. Use `console.log`, `codex.tool(...)`, and `codex.emitImage(...)`.";
@@ -194,7 +194,7 @@ async fn js_repl_tools_only_instructions_are_feature_gated() {
         .set(features)
         .expect("test config should allow js_repl tool restrictions");
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("js_repl instructions expected");
     let expected = "## JavaScript REPL (Node)\n- Use `js_repl` for Node-backed JavaScript with top-level await in a persistent kernel.\n- `js_repl` is a freeform/custom tool. Direct `js_repl` calls must send raw JavaScript tool input (optionally with first-line `// codex-js-repl: timeout_ms=15000`). Do not wrap code in JSON (for example `{\"code\":\"...\"}`), quotes, or markdown code fences.\n- Helpers: `codex.cwd`, `codex.homeDir`, `codex.tmpDir`, `codex.tool(name, args?)`, and `codex.emitImage(imageLike)`.\n- `codex.tool` executes a normal tool call and resolves to the raw tool output object. Use it for shell and non-shell tools alike. Nested tool outputs stay inside JavaScript unless you emit them explicitly.\n- `codex.emitImage(...)` adds one image to the outer `js_repl` function output each time you call it, so you can call it multiple times to emit multiple images. It accepts a data URL, a single `input_image` item, an object like `{ bytes, mimeType }`, or a raw tool response object with exactly one image and no text. It rejects mixed text-and-image content.\n- `codex.tool(...)` and `codex.emitImage(...)` keep stable helper identities across cells. Saved references and persisted objects can reuse them in later cells, but async callbacks that fire after a cell finishes still fail because no exec is active.\n- Request full-resolution image processing with `detail: \"original\"` only when the `view_image` tool schema includes a `detail` argument. The same availability applies to `codex.emitImage(...)`: if `view_image.detail` is present, you may also pass `detail: \"original\"` there. Use this when high-fidelity image perception or precise localization is needed, especially for CUA agents.\n- Example of sharing an in-memory Playwright screenshot: `await codex.emitImage({ bytes: await page.screenshot({ type: \"jpeg\", quality: 85 }), mimeType: \"image/jpeg\", detail: \"original\" })`.\n- Example of sharing a local image tool result: `await codex.emitImage(codex.tool(\"view_image\", { path: \"/absolute/path\", detail: \"original\" }))`.\n- When encoding an image to send with `codex.emitImage(...)` or `view_image`, prefer JPEG at about 85 quality when lossy compression is acceptable; use PNG when transparency or lossless detail matters. Smaller uploads are faster and less likely to hit size limits.\n- Top-level bindings persist across cells. If a cell throws, prior bindings remain available and bindings that finished initializing before the throw often remain usable in later cells. For code you plan to reuse across cells, prefer declaring or assigning it in direct top-level statements before operations that might throw. If you hit `SyntaxError: Identifier 'x' has already been declared`, first reuse the existing binding, reassign a previously declared `let`, or pick a new descriptive name. Use `{ ... }` only for a short temporary block when you specifically need local scratch names; do not wrap an entire cell in block scope if you want those names reusable later. Reset the kernel with `js_repl_reset` only when you need a clean state.\n- Top-level static import declarations (for example `import x from \"./file.js\"`) are currently unsupported in `js_repl`; use dynamic imports with `await import(\"pkg\")`, `await import(\"./file.js\")`, or `await import(\"/abs/path/file.mjs\")` instead. Imported local files must be ESM `.js`/`.mjs` files and run in the same REPL VM context. Bare package imports always resolve from REPL-global search roots (`CODEX_JS_REPL_NODE_MODULE_DIRS`, then cwd), not relative to the imported file location. Local files may statically import only other local relative/absolute/`file://` `.js`/`.mjs` files; package and builtin imports from local files must stay dynamic. `import.meta.resolve()` returns importable strings such as `file://...`, bare package names, and `node:...` specifiers. Local file modules reload between execs, while top-level bindings persist until `js_repl_reset`.\n- Do not call tools directly; use `js_repl` + `codex.tool(...)` for all tool calls, including shell commands.\n- MCP tools (if any) can also be called by name via `codex.tool(...)`.\n- Avoid direct access to `process.stdout` / `process.stderr` / `process.stdin`; it can corrupt the JSON line protocol. Use `console.log`, `codex.tool(...)`, and `codex.emitImage(...)`.";
@@ -213,7 +213,7 @@ async fn js_repl_image_detail_original_does_not_change_instructions() {
         .set(features)
         .expect("test config should allow js_repl image detail settings");
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("js_repl instructions expected");
     let expected = "## JavaScript REPL (Node)\n- Use `js_repl` for Node-backed JavaScript with top-level await in a persistent kernel.\n- `js_repl` is a freeform/custom tool. Direct `js_repl` calls must send raw JavaScript tool input (optionally with first-line `// codex-js-repl: timeout_ms=15000`). Do not wrap code in JSON (for example `{\"code\":\"...\"}`), quotes, or markdown code fences.\n- Helpers: `codex.cwd`, `codex.homeDir`, `codex.tmpDir`, `codex.tool(name, args?)`, and `codex.emitImage(imageLike)`.\n- `codex.tool` executes a normal tool call and resolves to the raw tool output object. Use it for shell and non-shell tools alike. Nested tool outputs stay inside JavaScript unless you emit them explicitly.\n- `codex.emitImage(...)` adds one image to the outer `js_repl` function output each time you call it, so you can call it multiple times to emit multiple images. It accepts a data URL, a single `input_image` item, an object like `{ bytes, mimeType }`, or a raw tool response object with exactly one image and no text. It rejects mixed text-and-image content.\n- `codex.tool(...)` and `codex.emitImage(...)` keep stable helper identities across cells. Saved references and persisted objects can reuse them in later cells, but async callbacks that fire after a cell finishes still fail because no exec is active.\n- Request full-resolution image processing with `detail: \"original\"` only when the `view_image` tool schema includes a `detail` argument. The same availability applies to `codex.emitImage(...)`: if `view_image.detail` is present, you may also pass `detail: \"original\"` there. Use this when high-fidelity image perception or precise localization is needed, especially for CUA agents.\n- Example of sharing an in-memory Playwright screenshot: `await codex.emitImage({ bytes: await page.screenshot({ type: \"jpeg\", quality: 85 }), mimeType: \"image/jpeg\", detail: \"original\" })`.\n- Example of sharing a local image tool result: `await codex.emitImage(codex.tool(\"view_image\", { path: \"/absolute/path\", detail: \"original\" }))`.\n- When encoding an image to send with `codex.emitImage(...)` or `view_image`, prefer JPEG at about 85 quality when lossy compression is acceptable; use PNG when transparency or lossless detail matters. Smaller uploads are faster and less likely to hit size limits.\n- Top-level bindings persist across cells. If a cell throws, prior bindings remain available and bindings that finished initializing before the throw often remain usable in later cells. For code you plan to reuse across cells, prefer declaring or assigning it in direct top-level statements before operations that might throw. If you hit `SyntaxError: Identifier 'x' has already been declared`, first reuse the existing binding, reassign a previously declared `let`, or pick a new descriptive name. Use `{ ... }` only for a short temporary block when you specifically need local scratch names; do not wrap an entire cell in block scope if you want those names reusable later. Reset the kernel with `js_repl_reset` only when you need a clean state.\n- Top-level static import declarations (for example `import x from \"./file.js\"`) are currently unsupported in `js_repl`; use dynamic imports with `await import(\"pkg\")`, `await import(\"./file.js\")`, or `await import(\"/abs/path/file.mjs\")` instead. Imported local files must be ESM `.js`/`.mjs` files and run in the same REPL VM context. Bare package imports always resolve from REPL-global search roots (`CODEX_JS_REPL_NODE_MODULE_DIRS`, then cwd), not relative to the imported file location. Local files may statically import only other local relative/absolute/`file://` `.js`/`.mjs` files; package and builtin imports from local files must stay dynamic. `import.meta.resolve()` returns importable strings such as `file://...`, bare package names, and `node:...` specifiers. Local file modules reload between execs, while top-level bindings persist until `js_repl_reset`.\n- Avoid direct access to `process.stdout` / `process.stderr` / `process.stdin`; it can corrupt the JSON line protocol. Use `console.log`, `codex.tool(...)`, and `codex.emitImage(...)`.";
@@ -229,13 +229,9 @@ async fn merges_existing_instructions_with_project_doc() {
 
     const INSTRUCTIONS: &str = "base instructions";
 
-    let res = get_user_instructions(
-        &make_config(&tmp, 4096, Some(INSTRUCTIONS)).await,
-        None,
-        None,
-    )
-    .await
-    .expect("should produce a combined instruction string");
+    let res = get_user_instructions(&make_config(&tmp, 4096, Some(INSTRUCTIONS)).await, None)
+        .await
+        .expect("should produce a combined instruction string");
 
     let expected = format!("{INSTRUCTIONS}{PROJECT_DOC_SEPARATOR}{}", "proj doc");
 
@@ -250,12 +246,7 @@ async fn keeps_existing_instructions_when_doc_missing() {
 
     const INSTRUCTIONS: &str = "some instructions";
 
-    let res = get_user_instructions(
-        &make_config(&tmp, 4096, Some(INSTRUCTIONS)).await,
-        None,
-        None,
-    )
-    .await;
+    let res = get_user_instructions(&make_config(&tmp, 4096, Some(INSTRUCTIONS)).await, None).await;
 
     assert_eq!(res, Some(INSTRUCTIONS.to_string()));
 }
@@ -284,7 +275,7 @@ async fn concatenates_root_and_cwd_docs() {
     let mut cfg = make_config(&repo, 4096, None).await;
     cfg.cwd = nested;
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("doc expected");
     assert_eq!(res, "root doc\n\ncrate doc");
@@ -312,7 +303,7 @@ async fn project_root_markers_are_honored_for_agents_discovery() {
     assert_eq!(discovery[0], expected_parent);
     assert_eq!(discovery[1], expected_child);
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("doc expected");
     assert_eq!(res, "parent doc\n\nchild doc");
@@ -327,7 +318,7 @@ async fn agents_local_md_preferred() {
 
     let cfg = make_config(&tmp, 4096, None).await;
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("local doc expected");
 
@@ -349,7 +340,7 @@ async fn uses_configured_fallback_when_agents_missing() {
 
     let cfg = make_config_with_fallback(&tmp, 4096, None, &["EXAMPLE.md"]).await;
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("fallback doc expected");
 
@@ -365,7 +356,7 @@ async fn agents_md_preferred_over_fallbacks() {
 
     let cfg = make_config_with_fallback(&tmp, 4096, None, &["EXAMPLE.md", ".example.md"]).await;
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("AGENTS.md should win");
 
@@ -398,7 +389,6 @@ async fn skills_are_appended_to_project_doc() {
     let res = get_user_instructions(
         &cfg,
         skills.errors.is_empty().then_some(skills.skills.as_slice()),
-        None,
     )
     .await
     .expect("instructions expected");
@@ -426,7 +416,6 @@ async fn skills_render_without_project_doc() {
     let res = get_user_instructions(
         &cfg,
         skills.errors.is_empty().then_some(skills.skills.as_slice()),
-        None,
     )
     .await
     .expect("instructions expected");
@@ -449,7 +438,7 @@ async fn apps_feature_does_not_emit_user_instructions_by_itself() {
         .enable(Feature::Apps)
         .expect("test config should allow apps");
 
-    let res = get_user_instructions(&cfg, None, None).await;
+    let res = get_user_instructions(&cfg, None).await;
     assert_eq!(res, None);
 }
 
@@ -463,7 +452,7 @@ async fn apps_feature_does_not_append_to_project_doc_user_instructions() {
         .enable(Feature::Apps)
         .expect("test config should allow apps");
 
-    let res = get_user_instructions(&cfg, None, None)
+    let res = get_user_instructions(&cfg, None)
         .await
         .expect("instructions expected");
     assert_eq!(res, "base doc");
